refactor: move files and update hooks
