+++ operate/client/src/modules/api/v2/processes/resolveProcessInstancesIncidentsBatchOperation.ts
